updating files for cmake. now presented with dll error for OpenAL. Preferablly we static link these libraries, will require digging through their cmakelists to see if that is an option / figuring out how to specify it via cmake. Old visual studio project still compiling.

+++ Space_Battle_Arcade/Space_Battle_Arcade/Source/GameFramework/SAPlayerSystem.h
@@ -24,7 +24,7 @@ namespace SA
 		{
 			static_assert(std::is_base_of<SA::PlayerBase, T>::value, "Players must be of the base class PlayerBase.");
 
-			sp<T> newPlayer = new_sp<T>(players.size());
+			sp<T> newPlayer = new_sp<T>(static_cast<int32_t>(players.size()));
 			players.push_back(newPlayer);
 
 			//broadcast that player was created with explicit PlayerBase type
